Enhance figure management: update GitHub Actions workflow to include specific figure names and add functions to retrieve figure name mappings in the analysis module.
