+++ .github/workflows/create-release.yaml
@@ -79,9 +79,11 @@ jobs:
       matrix:
         python-version: ["3.12"]
         figure-name:
-          - figure1
-          - figure2
-          - figure3 # Replace with actual figure names
+          - First 30 Days Active by Final Result
+          - Top Activities Total Clicks
+          - EDM LA Venn
+          - AutoML Feature Engineering
+          - Critical Difference Nemenyi
     steps:
       - uses: actions/checkout@v4
         with:
@@ -134,7 +136,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: figure-${{ matrix.figure-name }}
-          path: research/figures/generated/${{ matrix.figure-name }}
+          path: research/figures/generated/*.png
 
   build-pdf:
     runs-on: ubuntu-latest
