+++ src/packages/analysis/src/analysis/__main__.py
@@ -1,24 +1,7 @@
 from click import group, option
 from tqdm import tqdm
 
-from analysis.get_figures import (
-    PaperFigures,
-    PresentationFigures,
-    ProjectFigure,
-    SharedFigures,
-)
-
-
-def get_figure_name_map(
-    paper: bool = True, presentation: bool = True
-) -> dict[str, ProjectFigure]:
-    """Get a map of figure names to figure objects."""
-    figmap = {x.value.name: x.value for x in SharedFigures}
-    if paper:
-        figmap.update({x.value.name: x.value for x in PaperFigures})
-    if presentation:
-        figmap.update({x.value.name: x.value for x in PresentationFigures})
-    return figmap
+from analysis import get_figure_name_map
 
 
 @group
